Move copied files to externals dir
