+++ externals/TransportConnection.Generated.cs
@@ -1,115 +1,15 @@
-﻿using System;
-using System.Buffers;
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.IO.Pipelines;
-using System.Net;
-using System.Threading;
-using Microsoft.AspNetCore.Connections;
+
 using Microsoft.AspNetCore.Connections.Features;
 using Microsoft.AspNetCore.Http.Features;
 
-namespace RedHat.AspNetCore.Server.Kestrel.Transport.Linux
+namespace Microsoft.AspNetCore.Connections
 {
-    // Copied from https://github.com/aspnet/AspNetCore/tree/e0e9096af526448cb1a02f97e60efd8567b7ba35
-    internal abstract partial class TransportConnection : ConnectionContext
-    {
-        private IDictionary<object, object> _items;
-        private string _connectionId;
-
-        public TransportConnection()
-        {
-            FastReset();
-        }
-
-        public override EndPoint LocalEndPoint { get; set; }
-        public override EndPoint RemoteEndPoint { get; set; }
-
-        public override string ConnectionId
-        {
-            get
-            {
-                if (_connectionId == null)
-                {
-                    _connectionId = CorrelationIdGenerator.GetNextId();
-                }
-
-                return _connectionId;
-            }
-            set
-            {
-                _connectionId = value;
-            }
-        }
-
-        public override IFeatureCollection Features => this;
-
-        public virtual MemoryPool<byte> MemoryPool { get; }
-
-        public override IDuplexPipe Transport { get; set; }
-
-        public IDuplexPipe Application { get; set; }
-
-        public override IDictionary<object, object> Items
-        {
-            get
-            {
-                // Lazily allocate connection metadata
-                return _items ?? (_items = new ConnectionItems());
-            }
-            set
-            {
-                _items = value;
-            }
-        }
-
-        public override CancellationToken ConnectionClosed { get; set; }
-
-        // DO NOT remove this override to ConnectionContext.Abort. Doing so would cause
-        // any TransportConnection that does not override Abort or calls base.Abort
-        // to stack overflow when IConnectionLifetimeFeature.Abort() is called.
-        // That said, all derived types should override this method should override
-        // this implementation of Abort because canceling pending output reads is not
-        // sufficient to abort the connection if there is backpressure.
-        public override void Abort(ConnectionAbortedException abortReason)
-        {
-            Application.Input.CancelPendingRead();
-        }
-    }
-
-    internal partial class TransportConnection : IConnectionIdFeature,
-                                                 IConnectionTransportFeature,
-                                                 IConnectionItemsFeature,
-                                                 IMemoryPoolFeature,
-                                                 IConnectionLifetimeFeature
-    {
-        // NOTE: When feature interfaces are added to or removed from this TransportConnection class implementation,
-        // then the list of `features` in the generated code project MUST also be updated.
-        // See also: tools/CodeGenerator/TransportConnectionFeatureCollection.cs
-
-        MemoryPool<byte> IMemoryPoolFeature.MemoryPool => MemoryPool;
-
-        IDuplexPipe IConnectionTransportFeature.Transport
-        {
-            get => Transport;
-            set => Transport = value;
-        }
-
-        IDictionary<object, object> IConnectionItemsFeature.Items
-        {
-            get => Items;
-            set => Items = value;
-        }
-
-        CancellationToken IConnectionLifetimeFeature.ConnectionClosed
-        {
-            get => ConnectionClosed;
-            set => ConnectionClosed = value;
-        }
-
-        void IConnectionLifetimeFeature.Abort() => Abort(new ConnectionAbortedException("The connection was aborted by the application via IConnectionLifetimeFeature.Abort()."));
-    }
-
     internal partial class TransportConnection : IFeatureCollection
     {
         private static readonly Type IConnectionIdFeatureType = typeof(IConnectionIdFeature);
@@ -344,40 +244,5 @@ namespace RedHat.AspNetCore.Server.Kestrel.Transport.Linux
         IEnumerator<KeyValuePair<Type, object>> IEnumerable<KeyValuePair<Type, object>>.GetEnumerator() => FastEnumerable().GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => FastEnumerable().GetEnumerator();
-    }
-
-    internal static class CorrelationIdGenerator
-    {
-        // Base32 encoding - in ascii sort order for easy text based sorting
-        private static readonly char[] s_encode32Chars = "0123456789ABCDEFGHIJKLMNOPQRSTUV".ToCharArray();
-
-        // Seed the _lastConnectionId for this application instance with
-        // the number of 100-nanosecond intervals that have elapsed since 12:00:00 midnight, January 1, 0001
-        // for a roughly increasing _lastId over restarts
-        private static long _lastId = DateTime.UtcNow.Ticks;
-
-        public static string GetNextId() => GenerateId(Interlocked.Increment(ref _lastId));
-
-        private static string GenerateId(long id)
-        {
-            return string.Create(13, id, (buffer, value) =>
-            {
-                char[] encode32Chars = s_encode32Chars;
-
-                buffer[12] = encode32Chars[value & 31];
-                buffer[11] = encode32Chars[(value >> 5) & 31];
-                buffer[10] = encode32Chars[(value >> 10) & 31];
-                buffer[9] = encode32Chars[(value >> 15) & 31];
-                buffer[8] = encode32Chars[(value >> 20) & 31];
-                buffer[7] = encode32Chars[(value >> 25) & 31];
-                buffer[6] = encode32Chars[(value >> 30) & 31];
-                buffer[5] = encode32Chars[(value >> 35) & 31];
-                buffer[4] = encode32Chars[(value >> 40) & 31];
-                buffer[3] = encode32Chars[(value >> 45) & 31];
-                buffer[2] = encode32Chars[(value >> 50) & 31];
-                buffer[1] = encode32Chars[(value >> 55) & 31];
-                buffer[0] = encode32Chars[(value >> 60) & 31];
-            });
-        }
     }
 }
